Move card actions to the top of the popup
When saving a page, the card popup gets populated with data received
from the Wallabag API. The title and the image of the page then cause
the action bar to move.

By putting the action bar at the top of the popup, instead of on the
bottom, it does not move any more when the popup gets updated.

Signed-off-by: Jacob Kiers <jacob@jacobkiers.net>